requirements.txt
pyside
sqlalchemy
sqlite3
